Added quick implementation of sleep code
This is just one implementation of how we can put the board to sleep. This way shuts down every module except the bare necessity to keep track of how long the board will be put to sleep. Looking at the Microchip documentation if I was reading it right the power saving modes in the AVR_IoT_Cellular library it doesn't shut down every module and would leave some up like the modem? I could have just read it wrong idk will figure it out asking ross tomorrow.

The code is rather simple too, we just edit one variable to set how long we want the board to be asleep for in seconds. Read the big comment wall I left though to get the code to actually work, there is a slight hiccup in having too many definitions for an interrupt across all the various header files, annoying :P